LinkedIn-Scraping/requirements.txt
selenium==4.38.0
python-dotenv==1.2.1
pandas==2.3.3
streamlit==1.51.0